Cleanup instance imports
